[ONT-37] Moved trait detail panel
