+++ src/components/trait/TraitDetailPanel.vue
